firewalldb: move bbolt actions code to dedicated file
In preparation for adding a SQL implementation of the the Actions DB, we
move all kvdb CRUD code to a dedicated file.
